the dummy renderer now at least builds
