+++ trapezoid-core/src/gpu.rs
@@ -1,29 +1,33 @@
 mod command;
-mod front_blit;
-mod gpu_backend;
-mod gpu_context;
+mod common;
+
+#[cfg(feature="vulkan")]
+pub mod vulkan;
+
+#[cfg(feature="vulkan")]
+pub use vulkan as backend;
+
+#[cfg(feature="dummy_render")]
+pub mod dummy_render;
+
+#[cfg(feature="dummy_render")]
+pub use dummy_render as backend;
 
 use crate::memory::{interrupts::InterruptRequester, BusLine, Result};
 use command::{instantiate_gp0_command, Gp0CmdType, Gp0Command};
-use gpu_backend::GpuBackend;
+use backend::GpuBackend;
+use backend::StandardCommandBufferAllocator;
 
 use crossbeam::{
     atomic::AtomicCell,
     channel::{Receiver, Sender},
 };
-use vulkano::{
-    command_buffer::{
-        allocator::StandardCommandBufferAllocator, AutoCommandBufferBuilder, BlitImageInfo,
-        CommandBufferUsage, PrimaryAutoCommandBuffer,
-    },
-    device::{Device, Queue},
-    image::{sampler::Filter, Image},
-    sync::GpuFuture,
-};
 
 use std::{ops::Range, sync::Arc, thread::JoinHandle};
 
-use self::gpu_context::{DrawingTextureParams, DrawingVertex};
+use common::{DrawingTextureParams, DrawingVertex};
+
+pub use backend::{Device, Queue, Image, GpuFuture, AutoCommandBufferBuilder, PrimaryAutoCommandBuffer, BlitImageInfo, Filter, CommandBufferUsage};
 
 bitflags::bitflags! {
     #[derive(Default, Clone, Copy, Debug, PartialEq, Eq)]
@@ -400,6 +404,17 @@ impl Gpu {
         self.in_vblank
     }
 
+    #[cfg(feature="dummy_render")]
+    pub fn sync_gpu_and_blit_to_front(
+        &mut self,
+        _dest_image: Arc<Image>,
+        _full_vram: bool,
+        in_future: Box<dyn GpuFuture>,
+    ) -> Box<dyn GpuFuture> {
+        in_future
+    }
+
+    #[cfg(feature="vulkan")]
     pub fn sync_gpu_and_blit_to_front(
         &mut self,
         dest_image: Arc<Image>,
