+++ trapezoid-core/src/gpu/command.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use crossbeam::atomic::AtomicCell;
 
-use super::gpu_context::{vertex_position_from_u32, DrawingTextureParams, DrawingVertex};
+use crate::gpu::common::{vertex_position_from_u32, DrawingTextureParams, DrawingVertex};
 use super::{BackendCommand, GpuStat, GpuStateSnapshot};
 
 #[derive(Debug)]
