+++ trapezoid-core/src/memory.rs
@@ -11,7 +11,7 @@ use std::path::Path;
 use std::sync::Arc;
 
 use byteorder::{ByteOrder, LittleEndian, ReadBytesExt};
-use vulkano::device::{Device, Queue};
+use crate::gpu::{Device, Queue};
 
 use crate::cdrom::Cdrom;
 use crate::controller_mem_card::ControllerAndMemoryCard;
