+++ trapezoid-core/src/vulkan_backend/gpu/shaders/vertex.glsl
